Rename packages in Java code
