+++ src/main/java/com/github/jtendermint/jabci/api/IDeliverTx.java
@@ -23,8 +23,8 @@
  */
 package com.github.jtendermint.jabci.api;
 
-import com.github.jtendermint.jabci.types.RequestDeliverTx;
-import com.github.jtendermint.jabci.types.ResponseDeliverTx;
+import com.github.jtendermint.jabci.proto.types.RequestDeliverTx;
+import com.github.jtendermint.jabci.proto.types.ResponseDeliverTx;
 
 public interface IDeliverTx {
