+++ src/main/java/com/github/jtendermint/jabci/api/IInitChain.java
@@ -23,8 +23,8 @@
  */
 package com.github.jtendermint.jabci.api;
 
-import com.github.jtendermint.jabci.types.RequestInitChain;
-import com.github.jtendermint.jabci.types.ResponseInitChain;
+import com.github.jtendermint.jabci.proto.types.RequestInitChain;
+import com.github.jtendermint.jabci.proto.types.ResponseInitChain;
 
 public interface IInitChain {
     /**
