+++ src/main/java/com/github/jtendermint/jabci/api/ISetOption.java
@@ -23,8 +23,8 @@
  */
 package com.github.jtendermint.jabci.api;
 
-import com.github.jtendermint.jabci.types.RequestSetOption;
-import com.github.jtendermint.jabci.types.ResponseSetOption;
+import com.github.jtendermint.jabci.proto.types.RequestSetOption;
+import com.github.jtendermint.jabci.proto.types.ResponseSetOption;
 
 public interface ISetOption {
