+++ src/main/java/com/github/jtendermint/jabci/socket/DefaultFallbackListener.java
@@ -28,28 +28,28 @@ import org.slf4j.LoggerFactory;
 
 import com.github.jtendermint.jabci.api.ABCIAPI;
 import com.github.jtendermint.jabci.api.CodeType;
-import com.github.jtendermint.jabci.types.RequestBeginBlock;
-import com.github.jtendermint.jabci.types.RequestCheckTx;
-import com.github.jtendermint.jabci.types.RequestCommit;
-import com.github.jtendermint.jabci.types.RequestDeliverTx;
-import com.github.jtendermint.jabci.types.RequestEcho;
-import com.github.jtendermint.jabci.types.RequestEndBlock;
-import com.github.jtendermint.jabci.types.RequestFlush;
-import com.github.jtendermint.jabci.types.RequestInfo;
-import com.github.jtendermint.jabci.types.RequestInitChain;
-import com.github.jtendermint.jabci.types.RequestQuery;
-import com.github.jtendermint.jabci.types.RequestSetOption;
-import com.github.jtendermint.jabci.types.ResponseBeginBlock;
-import com.github.jtendermint.jabci.types.ResponseCheckTx;
-import com.github.jtendermint.jabci.types.ResponseCommit;
-import com.github.jtendermint.jabci.types.ResponseDeliverTx;
-import com.github.jtendermint.jabci.types.ResponseEcho;
-import com.github.jtendermint.jabci.types.ResponseEndBlock;
-import com.github.jtendermint.jabci.types.ResponseFlush;
-import com.github.jtendermint.jabci.types.ResponseInfo;
-import com.github.jtendermint.jabci.types.ResponseInitChain;
-import com.github.jtendermint.jabci.types.ResponseQuery;
-import com.github.jtendermint.jabci.types.ResponseSetOption;
+import com.github.jtendermint.jabci.proto.types.RequestBeginBlock;
+import com.github.jtendermint.jabci.proto.types.RequestCheckTx;
+import com.github.jtendermint.jabci.proto.types.RequestCommit;
+import com.github.jtendermint.jabci.proto.types.RequestDeliverTx;
+import com.github.jtendermint.jabci.proto.types.RequestEcho;
+import com.github.jtendermint.jabci.proto.types.RequestEndBlock;
+import com.github.jtendermint.jabci.proto.types.RequestFlush;
+import com.github.jtendermint.jabci.proto.types.RequestInfo;
+import com.github.jtendermint.jabci.proto.types.RequestInitChain;
+import com.github.jtendermint.jabci.proto.types.RequestQuery;
+import com.github.jtendermint.jabci.proto.types.RequestSetOption;
+import com.github.jtendermint.jabci.proto.types.ResponseBeginBlock;
+import com.github.jtendermint.jabci.proto.types.ResponseCheckTx;
+import com.github.jtendermint.jabci.proto.types.ResponseCommit;
+import com.github.jtendermint.jabci.proto.types.ResponseDeliverTx;
+import com.github.jtendermint.jabci.proto.types.ResponseEcho;
+import com.github.jtendermint.jabci.proto.types.ResponseEndBlock;
+import com.github.jtendermint.jabci.proto.types.ResponseFlush;
+import com.github.jtendermint.jabci.proto.types.ResponseInfo;
+import com.github.jtendermint.jabci.proto.types.ResponseInitChain;
+import com.github.jtendermint.jabci.proto.types.ResponseQuery;
+import com.github.jtendermint.jabci.proto.types.ResponseSetOption;
 
 /**
  * The DefaultFallbackListener answers every incoming ABCI-request with a CodeType.OK
